Add multiple users
Added multiple users to the request generator. It will now generate requests for multiple users and increment the userid. It will send request 1 for all users and then move on to sending request 2 for all users and so on. I removed the sleep timer since I misunderstood the requirement about the time interval.

Resolves #3